Create MouseToScrollBar.lua
This is a simple Hammerspoon script to move the user's mouse from anywhere on the screen to the lower right side of the screen approximate to the scroll bar. 

The script triggers after the mouse has not moved for X amount of time 

There is a 0.5 second timer set in this script. You can change it as needed. 

The intended user is someone who uses an eye tracker to use the cursor rather than a physical mouse. The tracker does not always facilitate the movement of the mouse to the scroll bar. This script essentially forces the resting place of the mouse to be on the scroll bar, facilitating their needs.